# contracts/pg-target.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PG1 · PG2 목표 v1 확장 API",
    "version": "1.0.0",
    "description": "현재 2개 PG에 추가할 목표 계약. 기존 current 명세를 구현 상태의 기준으로 사용. 인증·merchant/receipt 조회·webhook outbox·fault injection은 미구현."
  },
...
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "PG1 / KAKAO"
    },
    {
      "url": "http://localhost:8091",
      "description": "PG2 / NAVER"
    }
  ],
  "paths": {
    "/pg/v1/payments/approve": {
      "post": {
        "operationId": "pgApprove",
        "summary": "pgApprove",
        "tags": [
          "현재 PG"
        ],
        "security": [
          {
            "serviceJWT": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "최신 PG payment snapshot. 최초 응답 캐시/불변 거래 영수증이 아님.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgPayment"
                },
                "example": {
                  "paymentId": "11111111-1111-4111-8111-111111111111",
                  "merchantTxId": "mp-11111111-1111-4111-8111-111111111111-11111111-1111-4111-8111-111111111111",
                  "currency": "KRW",
                  "amount": 10000,
                  "status": "APPROVED",
                  "refundedAmount": 0,
                  "simulated": true
                }
              }
            }
          },
          "400": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "401": {
            "description": "서비스 인증/권한 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "403": {
            "description": "서비스 인증/권한 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "404": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "409": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "500": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          }
        },
        "x-implementation": "planned-upgrade",
        "x-requirements": [
          "PG-01"
        ],
        "x-authorized-roles": [
          "LOCAL_NETWORK_ONLY"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PgApprove"
              },
              "example": {
                "merchantTxId": "mp-11111111-1111-4111-8111-111111111111-11111111-1111-4111-8111-111111111111",
                "currency": "KRW",
                "amount": 10000,
                "idempotencyKey": "sample"
              }
            }
          }
        },
        "x-service-scope": "pg:write"
      }
    },
    "/pg/v1/payments/{paymentId}/cancel": {
      "post": {
        "operationId": "pgCancel",
        "summary": "pgCancel",
        "tags": [
          "현재 PG"
        ],
        "security": [
          {
            "serviceJWT": []
          }
        ],
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "최신 PG payment snapshot. 최초 응답 캐시/불변 거래 영수증이 아님.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgPayment"
                },
                "example": {
                  "paymentId": "11111111-1111-4111-8111-111111111111",
                  "merchantTxId": "mp-11111111-1111-4111-8111-111111111111-11111111-1111-4111-8111-111111111111",
                  "currency": "KRW",
                  "amount": 10000,
                  "status": "APPROVED",
                  "refundedAmount": 0,
                  "simulated": true
                }
              }
            }
          },
          "400": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "401": {
            "description": "서비스 인증/권한 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "403": {
            "description": "서비스 인증/권한 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "404": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "409": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "500": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          }
        },
        "x-implementation": "planned-upgrade",
        "x-requirements": [
          "PG-01"
        ],
        "x-authorized-roles": [
          "LOCAL_NETWORK_ONLY"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PgCancel"
              },
              "example": {
                "idempotencyKey": "sample"
              }
            }
          }
        },
        "x-service-scope": "pg:write"
      }
    },
    "/pg/v1/payments/{paymentId}/refund": {
      "post": {
        "operationId": "pgRefund",
        "summary": "pgRefund",
        "tags": [
          "현재 PG"
        ],
        "security": [
          {
            "serviceJWT": []
          }
        ],
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "최신 PG payment snapshot. 최초 응답 캐시/불변 거래 영수증이 아님.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgPayment"
                },
                "example": {
                  "paymentId": "11111111-1111-4111-8111-111111111111",
                  "merchantTxId": "mp-11111111-1111-4111-8111-111111111111-11111111-1111-4111-8111-111111111111",
                  "currency": "KRW",
                  "amount": 10000,
                  "status": "APPROVED",
                  "refundedAmount": 0,
                  "simulated": true
                }
              }
            }
          },
          "400": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "401": {
            "description": "서비스 인증/권한 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "403": {
            "description": "서비스 인증/권한 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "404": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "409": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "500": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          }
        },
        "x-implementation": "planned-upgrade",
        "x-requirements": [
          "PG-01"
        ],
        "x-authorized-roles": [
          "LOCAL_NETWORK_ONLY"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PgRefund"
              },
              "example": {
                "amount": 1000,
                "idempotencyKey": "sample"
              }
            }
          }
        },
        "x-service-scope": "pg:write"
      }
    },
    "/pg/v1/payments/{paymentId}": {
      "get": {
        "operationId": "pgGetPayment",
        "summary": "pgGetPayment",
        "tags": [
          "현재 PG"
        ],
        "security": [
          {
            "serviceJWT": []
          }
        ],
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "최신 PG payment snapshot. 최초 응답 캐시/불변 거래 영수증이 아님.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgPayment"
                },
                "example": {
                  "paymentId": "11111111-1111-4111-8111-111111111111",
                  "merchantTxId": "mp-11111111-1111-4111-8111-111111111111-11111111-1111-4111-8111-111111111111",
                  "currency": "KRW",
                  "amount": 10000,
                  "status": "APPROVED",
                  "refundedAmount": 0,
                  "simulated": true
                }
              }
            }
          },
          "400": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "401": {
            "description": "서비스 인증/권한 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "403": {
            "description": "서비스 인증/권한 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "404": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "409": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          },
          "500": {
            "description": "provider별 오류 코드 KAKAO_PG_xxx/NAVER_PG_xxx 또는 HTTP_xxx",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgError"
                },
                "example": {
                  "code": "sample",
                  "message": "sample",
                  "timestamp": "2026-09-21T00:00:00Z",
                  "path": "sample"
                }
              }
            }
          }
        },
        "x-implementation": "planned-upgrade",
        "x-requirements": [
          "PG-01"
        ],
        "x-authorized-roles": [
          "LOCAL_NETWORK_ONLY"
        ],
        "x-service-scope": "pg:read"
      }
    },
...
  },
  "components": {
    "schemas": {
      "PgApprove": {
        "type": "object",
        "properties": {
          "merchantTxId": {
            "type": "string",
            "pattern": "\\S"
          },
          "currency": {
            "type": "string",
            "const": "KRW"
          },
          "amount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000000
          },
          "idempotencyKey": {
            "type": "string",
            "pattern": "\\S"
          }
        },
        "required": [
          "merchantTxId",
          "currency",
          "amount",
          "idempotencyKey"
        ]
      },
      "PgCancel": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "pattern": "\\S"
          }
        },
        "required": [
          "idempotencyKey"
        ]
      },
      "PgRefund": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000000
          },
          "idempotencyKey": {
            "type": "string",
            "pattern": "\\S"
          }
        },
        "required": [
          "amount",
          "idempotencyKey"
        ]
      },
      "PgPayment": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string",
            "format": "uuid"
          },
          "merchantTxId": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "const": "KRW"
          },
          "amount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000000
          },
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "CANCELLED",
              "PARTIALLY_REFUNDED",
              "REFUNDED"
            ]
          },
          "refundedAmount": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "simulated": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "paymentId",
          "merchantTxId",
          "currency",
          "amount",
          "status",
          "refundedAmount",
          "simulated"
        ]
      },
      "PgError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
...
    },
    "securitySchemes": {
      "serviceJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "RS256, allowlisted iss/sub, service audience, 60-second exp, operation scopes. Never browser credentials."
      }
    }
  }
}